requirements.txt
flask==3.0.3
flask-cors==4.0.1
sqlalchemy==2.0.32
pydantic==2.8.2
bleach==6.1.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
